+++ RELEASE_NOTES.md
@@ -1,4 +1,7 @@
 ## Version 0.1.0 - (tbd)
  * [Add rudimentary support for html components](https://github.com/plotly/Dash.NET/commit/42f846417cc811d379a0635493dc8c176a356297)
  * [Add possibility to add custom css/scripts to the dash app](https://github.com/plotly/Dash.NET/commit/0a812476afa0f31b9c698cd4ee1a1e3ff40a7a62)
- * Add Tabs and Tab core components
+ * [Add Tabs and Tab core components]()
+ * [Auto generate id for callback registration](https://github.com/plotly/Dash.NET/commit/0b9cd854b6c42463523199048130a729558ccbc5)
+ * [Prevent initial callback calls per default](https://github.com/plotly/Dash.NET/commit/a9e6c632dfa8736afb055a3340b6513490ea6d23)
+ * [Add Store core component](https://github.com/plotly/Dash.NET/commit/0b9cd854b6c42463523199048130a729558ccbc5)